# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: stylifier-api
host: mock.stylifier.com
basePath: /
schemes:
  - https
consumes:
  - application/json; charset=utf-8
  - application/json
paths:
  /register:
    x-swagger-router-controller: users
    post:
      description: create
      operationId: registerUser
      parameters:
        - name: userInfo
          in: body
          schema:
            $ref: '#/definitions/user_to_create'
      responses:
        "200":
          description: user created
          examples:
              application/json; charset=utf-8:
                {
                  jwt: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJGTHpoZVhFUUc1MVBPeWphV2wwSEllUTVvNWxKYk5mQSJ9.wSqyHgqfwhFey7CirsxCdu9JVMlN8qxLK1oxepFWWfo',
                }
...
  /brands:
    x-swagger-router-controller: users
    get:
      description: get
      operationId: getBrands
      parameters:
        - name: q
          in: query
          description: query value to search for, type specified with `in` parameter
          type: string
          x-example: "acme"
        - name: pagination
          in: query
          type: number
      responses:
        "200":
          description: return `allow [ tags ]`
          schema:
            type: array
            items:
              $ref: '#/definitions/users'
          examples:
            application/json; charset=utf-8:
              {
                pagination: "",
                "data": [{
                  "id": "9210dce8-25da-4275-a73a-1383b9774253",
                  "username": "acme",
                  "full_name": "acme",
                  "profile_picture": "https://upload.wikimedia.org/wikipedia/commons/thumb/8/80/Logo_acme.svg/1000px-Logo_acme.svg.png",
                }]
              }
  /users:
    x-swagger-router-controller: users
    get:
      description: get
      operationId: getUsers
      parameters:
        - name: q
          in: query
          description: query value to search for, type specified with `in` parameter
          type: string
          x-example: "tom"
        - name: pagination
          in: query
          type: number
      responses:
        "200":
          description: return list of users
          schema:
            type: array
            items:
              $ref: '#/definitions/user_short'
          examples:
            application/json; charset=utf-8:
              {
                pagination: "",
                "data": [{
                  "id": "9210dce8-25da-4275-a73a-1383b9774253",
                  "username": "mockuser",
                  "full_name": "mock user",
                  "profile_picture": "https://mock.stylifier.com/images/9210dce8-25da-4275-a73a-1383b9774253.jpeg",
                }]
              }
  '/users/{username}':
    x-swagger-router-controller: users
    get:
      description: get all user images
      operationId: getUserByUsername
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        "200":
          description: array of users
          schema:
            $ref: '#/definitions/user'
          examples:
            application/json; charset=utf-8:
              {
                "id": "9210dce8-25da-4275-a73a-1383b9774253",
                "username": "mockuser",
                "full_name": "mock user",
                "profile_picture": "https://mock.stylifier.com/images/9210dce8-25da-4275-a73a-1383b9774253.jpeg",
                "bio": "Praesent mauris mauris ipsum.",
                "counts": {
                  "follows": 10,
                  "followed_by": 2,
                  "media": 2,
                  "thread_asked_by": 1,
                  "thread_asked_by_open": 0,
                  "sponsored": 0
                },
                "rating": 4,
                "is_instagram_user": false,
                "is_brand": false,
                "styles": ["formal"],
                "sponsors": [{
                  "id": "9210dce8-25da-4275-a73a-1383b9774253",
                  "username": "acme",
                  "full_name": "acme",
                  "profile_picture": "https://upload.wikimedia.org/wikipedia/commons/thumb/8/80/Logo_acme.svg/1000px-Logo_acme.svg.png"
                }],
                "created_time": "1515094955466"
              }
...
  '/users/{username}/approve':
    x-swagger-router-controller: users
    post:
      description: approve a user for beta testing
      operationId: approveUser
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        "200":
          description: success true
          examples:
            application/json; charset=utf-8:
              {
                "success": true
              }
...
  '/user/self/request_approval':
    x-swagger-router-controller: users
    post:
      description: request for approval for beta testing
      operationId: requestApproveUser
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object

      responses:
        "200":
          description: success true
          examples:
            application/json; charset=utf-8:
              {
                "success": true
              }
  '/user/self/profile_picture':
    x-swagger-router-controller: users
    post:
      description: set users profile picture
      operationId: setProfilePicture
      parameters:
        - name: media
          in: body
          required: true
          schema:
            $ref: '#/definitions/media'

      responses:
        "200":
          description: profile picture has changed
          examples:
            application/json; charset=utf-8:
              {
                "success": true
              }
...
  '/media':
    x-swagger-router-controller: media

    get:
      description: search media
      operationId: getMedia
      parameters:
        - name: pagination
          in: query
          type: number
        - name: q
          in: query
          type: string
      responses:
        "200":
          description: array of links to images
          schema:
            $ref: '#/definitions/media_list'
          examples:
            application/json; charset=utf-8:
              {
                "pagination": "",
                "data": [{
                  "id": "d42e0091-2c67-437b-987c-296661c0dc3e",
                  "user": {
                    "id": "9210dce8-25da-4275-a73a-1383b9774253",
                    "username": "acme",
                    "full_name": "acme",
                    "profile_picture": "https://mock.stylifier.com/images/9210dce8-25da-4275-a73a-1383b9774253.jpeg"
                  },
                  "images": {
                    "thumbnail" : {
                      "url": "https://mock.stylifier.com/images/d42e0091-2c67-437b-987c-296661c0dc3e.jpeg",
                      "width": 640,
                      "height": 480
                    },
                    "low_resolution": {
                      "url": "https://mock.stylifier.com/images/d42e0091-2c67-437b-987c-296661c0dc3e.jpeg",
                      "width": 640,
                      "height": 480
                    },
                    "standard_resolution": {
                      "url": "https://mock.stylifier.com/images/d42e0091-2c67-437b-987c-296661c0dc3e.jpeg",
                      "width": 640,
                      "height": 480
                    }
                  },
                  "user_has_liked": false,
                  "likes": 10,
                  "type": image,
                  "users_in_photo": [],
                  "created_time": "2017-12-26T15:54:24.921Z",
                  "location": {
                    "latitude": 60.167,
                    "longitude": 24.955,
                    "name": "Market Square, Helsinki",
                    "id": "504979721",
                  }
                }]
              }

    post:
      description: upload new image
      operationId: createMedia
      consumes:
        - multipart/form-data
      responses:
        "200":
          description: Object of the image
          schema:
            $ref: '#/definitions/image'
          examples:
            application/json; charset=utf-8:
              {
                "success": true,
                "id": "1210dce8-25da-4275-a73a-1383b9774255"
              }
...
definitions:
  id:
    type: string
...
  user_to_create:
    type: object
    allOf:
      - $ref: '#/definitions/user_short'
      - type: object
        properties:
          email:
            type: string
          password:
            type: string
          instagram_code:
            type: string
  style:
    type: string
  user_short:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      username:
        type: string
      full_name:
        type: string
      profile_picture:
        type: string
  user:
    allOf:
      - $ref: '#/definitions/user_short'
      - type: object
        properties:
          bio:
            type: string
          counts:
            type: object
            properties:
              follows:
                type: number
              followed_by:
                type: number
              media:
                type: number
              thread_asked_by:
                type: number
              thread_asked_by_open:
                type: number
              sponsored:
                type: number
          contribution_earned:
            type: number
          rating:
            type: number
          request_rating:
            type: number
          is_instagram_user:
            type: boolean
          is_brand:
            type: boolean
          styles:
            type: array
            items:
              $ref: '#/definitions/style'
          sponsors:
            type: array
            items:
              $ref: '#/definitions/user_short'
          created_time:
            type: string
  users:
    type: object
    properties:
      pagination:
        type: string
      data:
        type: array
        items:
          $ref: '#/definitions/user'
...
  media:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      user:
        $ref: '#/definitions/user_short'
      images:
        $ref: '#/definitions/image'
      user_has_liked:
        type: boolean
      likes:
        type: number
      type:
        type: string
        enum:
          - image
      users_in_photo:
        type: array
        items:
          $ref: '#/definitions/user'
      created_time:
        type: string
...
  image:
    type: object
    properties:
      width:
        type: number
      height:
        type: number
      url:
        type: string
...
  media_list:
    type: object
    properties:
      pagination:
        type: string
      data:
        type: array
        items:
          $ref: '#/definitions/media'
...